Move global sport types to regular `.ts` file
