+++ dotcom-rendering/src/components/GetCricketScoreboard.importable.tsx
@@ -1,6 +1,7 @@
 import { ArticleDesign, type ArticleFormat } from '@guardian/libs';
 import type { SWRConfiguration } from 'swr';
 import { useApi } from '../lib/useApi';
+import type { CricketMatch } from '../types/sport';
 import { CricketScoreboard } from './CricketScoreboard';
 import { Placeholder } from './Placeholder';
