+++ dotcom-rendering/src/components/GetMatchNav.importable.tsx
@@ -3,6 +3,7 @@ import { ArticleDesign, type ArticleFormat } from '@guardian/libs';
 import { from } from '@guardian/source/foundations';
 import type { SWRConfiguration } from 'swr';
 import { useApi } from '../lib/useApi';
+import type { TeamType } from '../types/sport';
 import type { TagType } from '../types/tag';
 import { ArticleHeadline } from './ArticleHeadline';
 import { MatchNav } from './MatchNav';
